% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{cases}
\alias{cases}
\alias{cases-package}
\title{lettercase}
\description{
A package for formatting and manipulating various types of cases. This
package provides utilities that handles several properties.
}
\details{
* capitalization
* whitespace
* special characters
* acronyms
}
\references{
\url{http://en.wikipedia.org/wiki/Letter_case}
}

