% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_str_replace.R
\name{make_str_replace}
\alias{make_str_delete}
\alias{make_str_replace}
\title{Make replace, delete and "is" functions for strings}
\usage{
make_str_replace(pattern, replacement)

make_str_delete(pattern)
}
\arguments{
\item{pattern}{pattern to look for, as defined by a POSIX regular expression.
See the "Extended Regular Expressions" section of regex for details. See
fixed, ignore.case and perl for how to use other types of matching: fixed,
case insensitive and perl-compatible expressions.}

\item{replacement}{string. References of the form \code{\\1}, \code{\\2} will be replaced
with the contents of the respective matched group (created by ()) within the
pattern.

These functions build functions that take a single string argument and return
a vector as a result.

\code{make_str_replace} builds a functions that replaces the strings
according to the \code{pattern} and \code{replacement} arguments.

\code{make_str_delete} builds a functions that deletes the \code{pattern}
from the string.

\code{make_str_is} builds a function that detects is the string is has a
certain type of formatting.}
}
\description{
Functions for building string functions for replacement, deletion and testing
}
\examples{
# -tk
}

