% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\docType{data}
\name{pattern_whitespace}
\alias{pattern_whitespace}
\alias{pattern_whitespace_like}
\alias{pattern_separators}
\alias{pattern_ucfirst}
\alias{pattern_word}
\alias{pattern_nonword}
\alias{pattern_uppercase}
\alias{pattern_lowercase}
\title{Common regular expression patterns used by lettercase}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
pattern_whitespace

pattern_whitespace_like

pattern_separators

pattern_ucfirst

pattern_word

pattern_nonword

pattern_uppercase

pattern_lowercase
}
\description{
These are patterns that are used in the lettercase package.
}
\details{
\code{lettercase} provides a number of regular expression patterns that are
used by the case conversion functions.

These are not exported; to use them prefix them with \code{lettercase:::*}
}
\keyword{datasets}
