% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_cap_words.R
\name{str_cap_words}
\alias{cap_words}
\alias{str_cap_words}
\title{str_cap_words}
\usage{
str_cap_words(string)
}
\arguments{
\item{string}{character vector  to turn into a CapWords


CapWords is distinguished by:
\enumerate{
  \item All words with upper case first character
  \item No non-word characters allowed (letters/numbers only allowed)
}

The recipe for chaging into capwords is:
\enumerate{
  \item replace whitespace and other word separators with space
  \item str_ucfirst
  \item strstdelete whitespace
}}
}
\description{
Function used to convert character vectors to CapWord format.
}
\examples{
# CAP WORDS
    str_cap_words( "One Flew Over The Cuckoo's Nest" )
    str_cap_words( "Catch-22" )  # CATCH
}

