% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/str_delete.R
\name{str_delete}
\alias{str_delete}
\title{str_delete}
\usage{
str_delete(string, ...)
}
\arguments{
\item{string}{atomic character vector}

\item{...}{atomic patterns}
}
\description{
Delete or remove characters from a string based on one or more pattern
}
\examples{
str_delete( "ABC & 123", "\\\\W" )  # ABC123
}

