% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_sentence_case.R
\name{str_sentence_case}
\alias{str_sentence_case}
\title{str_sentence_case}
\usage{
str_sentence_case(string)
}
\arguments{
\item{string}{character string to transform

Sentance case is ...}
}
\description{
Format a string in sentence case
}

