% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/str_snake_case.R
\name{str_snake_case}
\alias{str_snake_case}
\title{str_snake_case}
\usage{
str_snake_case(string)
}
\arguments{
\item{string}{object to turn into a title case

* characters are all lower case
* non- \code{\\w}, \code{\\s} and - are dropped
* \code{\\w} and - are converted to underscore
* no support for acronyms}
}
\description{
Function used to convert character vectors to snake case format.
}
\examples{
str_snake_case( "One Flew Over The Cuckoo's Nest" )
 str_snake_case( "Catch-22" )  # catch_22
 str_snake_case( "Catch.22" )
}

