% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_title_case.R
\name{str_title_case}
\alias{str_title_case}
\title{str_title_case}
\usage{
str_title_case(x)
}
\arguments{
\item{x}{object to turn into a title case}
}
\description{
internal function for converting character to title case.  This is not
exported and should not be called directly.
}
\examples{
str_title_case( "One Flew Over The Cuckoo's Nest" )
  str_title_case( "one_flew_over_the_cuckoo_'_s_nest" )
}

