% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_ucfirst.R
\name{str_ucfirst}
\alias{is_ucfirst}
\alias{str_ucfirst}
\title{str_ucfirst}
\usage{
str_ucfirst(string)

is_ucfirst(string)
}
\arguments{
\item{string}{character vector to be converted.}
}
\description{
Convert first character of a string to uppercase
}
\examples{
str_ucfirst( "one flew over the cuckoo's nest" )
 str_ucfirst( "catch-22" )
 str_ucfirst( "Portrait of the Artist as a Young Man" )
# is_ucfirst
  is_ucfirst( 'ABC123' )      # TRUE
  is_ucfirst( 'abc123' )      # FALSE
  is_ucfirst( 'aBC'  )        # FALSE
  is_ucfirst( 'Abc' )         # TRUE
  is_ucfirst( 'Abc dEF' )     # FALSE
  is_ucfirst( '123' )         # TRUE
}
\seealso{
\code{\link{str_title_case}}
}

