% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_lowercase.R, R/str_uppercase.R
\name{str_lowercase}
\alias{is_all_caps}
\alias{is_allcaps}
\alias{is_lower}
\alias{is_lower_case}
\alias{is_lowercase}
\alias{is_upper}
\alias{is_upper_case}
\alias{is_uppercase}
\alias{lower_case}
\alias{lowercase}
\alias{str_all_caps}
\alias{str_allcaps}
\alias{str_capitalize}
\alias{str_decapitalize}
\alias{str_lower}
\alias{str_lower_case}
\alias{str_lowercase}
\alias{str_upper}
\alias{str_upper_case}
\alias{str_uppercase}
\alias{upper_case}
\alias{uppercase}
\title{str_uppercase, str_lowercases}
\usage{
str_lowercase(string)

str_lower_case(string)

str_lower(string)

str_decapitalize(string)

is_lowercase(string)

is_lower_case(string)

is_lower(string)

str_uppercase(string)

str_upper_case(string)

str_upper(string)

str_all_caps(string)

str_allcaps(string)

str_capitalize(string)

is_uppercase(string)

is_upper_case(string)

is_upper(string)

is_all_caps(string)

is_allcaps(string)
}
\arguments{
\item{string}{character; argument to be converted}
}
\description{
convert string to upper or lower case
}
\section{Upper Case and All Caps}{


  \strong{uppercase}, \strong{upper_case}, \strong{allcaps},
  \strong{all_caps} are all synonyms.

  \strong{upper case} or \code{all caps} contains all capital letters.
}

\section{Lower Case}{

  \code{str_all_caps} is a synonym for \code{str_uppercase}.
  \code{is_all_caps} is a synonym for \code{is_upper_case}.
}
\examples{
str_decapitalize( "ABC" )     # abc
# is_lower_case
  is_lowercase( 'ABC123' )      # FALSE
  is_lowercase( 'abc123' )      # TRUE
  is_lowercase( 'aB'  )         # FALSE
  is_lowercase( '123' )         # TRUE
# str_uppercase
  str_uppercase( "one flew over the cuckoo's nest" )
  str_uppercase( "catch-22" )

  str_capitalize( "abc" )               # ABC
  str_all_caps( "abc" )                 # ABC
# is_uppercase
  is_uppercase( 'ABC123' )      # TRUE
  is_uppercase( 'abc123' )      # FALSE
  is_uppercase( 'aB'  )         # FALSE
  is_uppercase( '123' )         # TRUE
}
\seealso{
\code{\link{str_is}}
}

