% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_delete.R
\name{str_delete}
\alias{str_delete}
\title{str_delete}
\usage{
str_delete(string, ...)
}
\arguments{
\item{string}{atomic character vector}

\item{...}{stringr-style matching \code{\link[stringr]{modifiers}}}
}
\description{
Delete or remove characters from a string based on one or more patterns
}
\details{
Deletes all occurences of the patterns from the string using
\code{\link[stringr]{str_replace_all}}
}
\examples{
str_delete( "ABC & 123", stringr::regex("\\\\W") )  # ABC123
}
\references{
\code{\link[stringr]{modifiers}} \cr
 \code{\link[stringr]{str_replace_all}}
}

